+++ packages/polythene-core-textfield/src/textfield.js
@@ -1,10 +1,8 @@
-import { filterSupportedAttributes, pointerStartEvent } from "polythene-core";
+import { filterSupportedAttributes } from "polythene-core";
 import classes from "polythene-css-classes/textfield";
 
 export const getElement = vnode =>
   vnode.attrs.element || "div";
-
-const MINIMUM_FOCUS_INTERVAL = 150;
 
 const DEFAULT_VALID_STATE = {
   invalid: false,
@@ -84,12 +82,13 @@ const notifyState = vnode => {
   if (attrs.onChange) {
     const status = getValidStatus(state, attrs);
     attrs.onChange({
-      focus:   state.hasFocus(),
-      dirty:   state.isDirty(),
-      el:      state.inputEl(),
-      invalid: status.invalid,
-      error:   status.error,
-      value:   state.inputEl().value,
+      focus:         state.hasFocus(),
+      dirty:         state.isDirty(),
+      el:            state.inputEl(),
+      invalid:       status.invalid,
+      error:         status.error,
+      value:         state.inputEl().value,
+      setInputState: newState => state.setInputState(newState),
     });
   }
 };
@@ -108,10 +107,9 @@ export const getInitialState = (vnode, createStream) => {
 
   const el = createStream(null);
   const inputEl = createStream(null);
-  const setValue = createStream({});
+  const setInputState = createStream({});
   const error = createStream(attrs.error);
-  const hasFocus = createStream(attrs.focus || false);
-  const setFocus = createStream(false);
+  const hasFocus = createStream(false);
   const isTouched = createStream(false); // true when any change is made
   const isDirty = createStream(defaultValue !== ""); // true for any input
   const isInvalid = createStream(false);
@@ -129,8 +127,7 @@ export const getInitialState = (vnode, createStream) => {
     isInvalid,
     isTouched,
     previousValue,
-    setFocus,
-    setValue,
+    setInputState,
     redrawOnUpdate: createStream.merge([inputEl, isInvalid, isDirty])
   };
 };
@@ -146,8 +143,9 @@ export const onMount = vnode => {
   vnode.state.inputEl(inputEl);
   state.inputEl().value = state.defaultValue;
 
-  state.setValue.map(({ type, focus }) => (
-    focus !== undefined && state.setFocus(focus),
+  state.setInputState.map(({ type, focus, value }) => (
+    value !== undefined ? state.inputEl().value = value : null,
+    focus !== undefined && (state.hasFocus(focus), focus ? state.inputEl().focus() : state.inputEl().blur()),
     type === "input" && (attrs.validateOnInput || attrs.counter) && state.isTouched(state.inputEl().value !== state.defaultValue),
     type !== "input" && state.isTouched(state.inputEl().value !== state.defaultValue),
     type === "onblur" && state.isTouched(true),
@@ -156,22 +154,6 @@ export const onMount = vnode => {
     notifyState(vnode),
     state.previousValue(state.inputEl().value)
   ));
-
-  state.setFocus.map(focusState => {
-    // Prevent autocomplete from getting in a loop
-    if ((state.didSetFocusTime + MINIMUM_FOCUS_INTERVAL) > new Date().getTime()) {
-      return;
-    }
-    state.hasFocus(focusState);
-    if (focusState && state.inputEl()) {
-      // Draw in next tick, to prevent getting an immediate onblur
-      // Explicit setting of focus needed for most browsers other than Safari
-      setTimeout(() => (
-        state.inputEl() && state.inputEl().focus && state.inputEl().focus(),
-        state.didSetFocusTime = new Date().getTime()
-      ), 0);
-    }
-  });
 
   notifyState(vnode);
 };
@@ -194,7 +176,7 @@ export const onUpdate = vnode => {
   if (inputEl && state.previousValue() !== valueStr) {
     inputEl.value = valueStr;
     state.previousValue(valueStr);
-    state.setValue({ type: "input" });
+    state.setInputState({ type: "input" });
   }
 };
 
@@ -248,10 +230,6 @@ export const createContent = (vnode, { renderer: h, keys: k }) => {
   const validates = !!(attrs.valid !== undefined || attrs.validate || attrs.min || attrs.max || attrs[k.minlength] || attrs[k.maxlength] || attrs.required || attrs.pattern);
   const inactive = attrs.disabled || attrs[k.readonly];
 
-  if (attrs.focus && !state.hasFocus() && !inactive) {
-    state.setFocus(true);
-  }
-
   const requiredIndicator = attrs.required && attrs.requiredIndicator !== ""
     ? h("span",
       {
@@ -286,15 +264,6 @@ export const createContent = (vnode, { renderer: h, keys: k }) => {
             {
               key: "label",
               className: classes.label,
-              // In IE10 the label catches click events on the field
-              // the function causes the input to get focus
-              [k["on" + pointerStartEvent]]: () => {
-                if (!inactive) {
-                  setTimeout(() => {
-                    state.inputEl.focus();
-                  }, 0);
-                }
-              }
             },
             label)
           : null,
@@ -318,7 +287,7 @@ export const createContent = (vnode, { renderer: h, keys: k }) => {
                 }
                 // in case the browser does not give the field focus,
                 // for instance when the user tapped to the current field off screen
-                state.setFocus(true);
+                state.setInputState({ focus: true });
                 notifyState(vnode);
               }
             }
@@ -330,7 +299,8 @@ export const createContent = (vnode, { renderer: h, keys: k }) => {
                 if (inactive) {
                   return;
                 }
-                state.setFocus(true);
+                state.setInputState({ focus: true });
+
                 // set CSS class manually in case field gets focus but is off screen
                 // and no redraw is triggered
                 // at the next redraw state.hasFocus() will be read and the focus class be set
@@ -346,7 +316,7 @@ export const createContent = (vnode, { renderer: h, keys: k }) => {
           !ignoreEvent(attrs, k.onblur)
             ? {
               [k.onblur]: () => {
-                state.setValue({ type: "onblur", focus: false });
+                state.setInputState({ type: "onblur", focus: false });
                 // same principle as onfocus
                 state.el().classList.remove(classes.stateFocused);
               }
@@ -358,7 +328,7 @@ export const createContent = (vnode, { renderer: h, keys: k }) => {
               [k.oninput]: () => {
                 // default input event
                 // may be overwritten by attrs.events
-                state.setValue({ type: "input" });
+                state.setInputState({ type: "input" });
               }
             }
             : null,
@@ -368,8 +338,8 @@ export const createContent = (vnode, { renderer: h, keys: k }) => {
               [k.onkeydown]: e => {
                 if (e.key === "Enter") {
                   state.isTouched(true);
-                } else if (e.key === "Escape") {
-                  inputEl.blur(e);
+                } else if (e.key === "Escape" || e.key === "Esc") {
+                  state.setInputState({ focus: false });
                 }
               }
             }